move gui parts
