+++ src/components/constant.rs
@@ -1,12 +1,6 @@
-use crate::{
-    common::{Component, ViziaComponent, Output, OutputType, Ports, Signal, Simulator},
-    gui_vizia::{popup::NewPopup, tooltip::new_component_tooltip},
-};
+use crate::common::{Component, Output, OutputType, Ports, Signal, Simulator};
 use serde::{Deserialize, Serialize};
-use vizia::{
-    prelude::*,
-    vg::{Paint, Path},
-};
+
 #[derive(Serialize, Deserialize)]
 pub struct Constant {
     pub id: String,
@@ -35,51 +29,5 @@ impl Component for Constant {
 
     fn evaluate(&self, simulator: &mut Simulator) {
         simulator.set_id_index(&self.id, 0, self.value);
-    }
-
-    
-}
-
-#[typetag::serde]
-impl ViziaComponent for Constant {
-    // create view
-    fn view(&self, cx: &mut Context) {
-        println!("---- Create Constant View");
-        View::build(ConstantView {}, cx, |cx| {
-            Label::new(cx, &format!("{:?}", self.value)).hoverable(false);
-            NewPopup::new(cx, self.get_id_ports()).position_type(PositionType::SelfDirected);
-        })
-        .position_type(PositionType::SelfDirected)
-        .left(Pixels(self.pos.0 - 10.0))
-        .top(Pixels(self.pos.1 - 10.0))
-        .width(Pixels(20.0))
-        .height(Pixels(20.0))
-        // TODO: do we want/need tooltip/popup for constants
-        .on_press(|ex| ex.emit(PopupEvent::Switch))
-        .tooltip(|cx| new_component_tooltip(cx, self));
-    }
-}
-pub struct ConstantView {}
-
-impl View for ConstantView {
-    fn element(&self) -> Option<&'static str> {
-        Some("Constant")
-    }
-
-    fn draw(&self, cx: &mut DrawContext<'_>, canvas: &mut Canvas) {
-        let bounds = cx.bounds();
-        // println!("Constant draw {:?}", bounds);
-
-        let mut path = Path::new();
-        let mut paint = Paint::color(vizia::vg::Color::rgbf(0.0, 1.0, 0.0));
-        paint.set_line_width(cx.logical_to_physical(1.0));
-
-        path.move_to(bounds.left() + 0.5, bounds.top() + 0.5);
-        path.line_to(bounds.right() + 0.5, bounds.top() + 0.5);
-        path.line_to(bounds.right() + 0.5, bounds.bottom() + 0.5);
-        path.line_to(bounds.left() + 0.5, bounds.bottom() + 0.5);
-        path.line_to(bounds.left() + 0.5, bounds.top() + 0.5);
-
-        canvas.fill_path(&path, &paint);
     }
 }
