+++ src/components/mux.rs
@@ -1,12 +1,5 @@
-use crate::{
-    common::{Component, ViziaComponent, Input, Output, OutputType, Ports, Simulator},
-    gui_vizia::{popup::NewPopup, tooltip::new_component_tooltip, GuiData},
-};
+use crate::common::{Component, Input, Output, OutputType, Ports, Simulator};
 use serde::{Deserialize, Serialize};
-use vizia::{
-    prelude::*,
-    vg::{Paint, Path},
-};
 
 #[derive(Serialize, Deserialize)]
 pub struct Mux {
@@ -46,90 +39,5 @@ impl Component for Mux {
 
         // set output
         simulator.set_id_index(&self.id, 0, value);
-    }
-
-}
-#[typetag::serde]
-impl ViziaComponent for Mux {
-    // create view
-    fn view(&self, cx: &mut Context) {
-        println!("---- Create Add View");
-
-        View::build(
-            MuxView {
-                select: self.select.clone(),
-            },
-            cx,
-            |cx| {
-                NewPopup::new(cx, self.get_id_ports()).position_type(PositionType::SelfDirected);
-            },
-        )
-        .position_type(PositionType::SelfDirected)
-        .left(Pixels(self.pos.0 - 20.0))
-        .top(Pixels(self.pos.1 - 10.0 * self.m_in.len() as f32 - 10.0))
-        .width(Pixels(40.0))
-        .height(Pixels(20.0 * self.m_in.len() as f32 + 20.0))
-        .on_press(|ex| ex.emit(PopupEvent::Switch))
-        .tooltip(|cx| new_component_tooltip(cx, self));
-    }
-}
-
-pub struct MuxView {
-    select: Input,
-}
-
-impl View for MuxView {
-    fn element(&self) -> Option<&'static str> {
-        Some("Mux")
-    }
-
-    fn draw(&self, cx: &mut DrawContext<'_>, canvas: &mut Canvas) {
-        let bounds = cx.bounds();
-        let scale = cx.scale_factor();
-        // println!("Mux draw {:?}", bounds);
-
-        let mut path = Path::new();
-        let mut paint = Paint::color(vizia::vg::Color::rgbf(0.0, 0.0, 0.0));
-        paint.set_line_width(cx.logical_to_physical(1.0));
-
-        let height = bounds.height();
-        let width = bounds.width();
-        let top = bounds.top();
-        let left = bounds.left();
-        let right = bounds.right();
-        let bottom = bounds.bottom();
-
-        // top left
-        path.move_to(left + 0.5, top + 0.5);
-
-        // top and right corner
-        path.line_to(left + width * 0.5 + 0.5, top + 0.5);
-        path.line_to(right + 0.5, top + height * 0.25 + 0.5);
-
-        // bottom and right corner
-        path.line_to(bounds.right() + 0.5, bottom - height * 0.25 + 0.5);
-        path.line_to(left + width * 0.5 + 0.5, bottom + 0.5);
-        path.line_to(left + 0.5, bottom + 0.5);
-
-        // left side
-        path.line_to(left + 0.5, top + 0.5);
-
-        canvas.stroke_path(&path, &paint);
-
-        // selector
-        let simulator = GuiData::simulator.get(cx);
-
-        let select = simulator.get_input_val(&self.select);
-
-        // println!("----- select = {}", select);
-        paint = Paint::color(vizia::vg::Color::rgbf(1.0, 0.0, 0.0));
-        let mut path = Path::new();
-
-        path.move_to(
-            left + 0.5,
-            top + 0.5 + (20.0 + select as f32 * 20.0) * scale,
-        );
-        path.line_to(right + 0.5, top + height * 0.5 + 0.5);
-        canvas.stroke_path(&path, &paint);
     }
 }
