+++ src/gui_vizia/components/mod.rs
@@ -1,3 +1,16 @@
 mod add;
+mod constant;
+mod mux;
+mod probe;
+mod probe_out;
+mod register;
+mod sext;
+mod wire;
 
-pub use add::*;
+// pub use add::*;
+// pub use constant::*;
+// pub use mux::*;
+// pub use probe::*;
+// pub use probe_out::*;
+// pub use sext::*;
+// pub use wire::*;
