+++ src/gui_vizia/components/probe_out.rs
@@ -1,0 +1,4 @@
+use crate::{common::ViziaComponent, components::ProbeOut};
+
+#[typetag::serde]
+impl ViziaComponent for ProbeOut {}
